The proxy functionality is now fully working.

+++ main.cpp
@@ -14,9 +14,9 @@ void parse_command_line(int argc, char **argv, po::variables_map& vm)
     po::options_description desc("Command line");
     desc.add_options()
     ("help", "Print usage information.")
-    ("lp", po::value<unsigned int>(), "The port on which to listen.")
+    ("lp", po::value<unsigned short>(), "The port on which to listen.")
     ("ra", po::value<string>(), "Address to redirect to.")
-    ("rp", po::value<unsigned int>(), "Port to redirect to.");
+    ("rp", po::value<unsigned short>(), "Port to redirect to.");
 
     po::store(po::parse_command_line(argc, argv, desc), vm);
     po::notify(vm);
@@ -49,7 +49,7 @@ int main(int argc, char **argv)
     try
     {
         boost::asio::io_service io_service;
-        server server(io_service, tcp::endpoint(tcp::v4(), vm["lp"].as<unsigned int>()), vm["ra"].as<string>(), vm["rp"].as<unsigned int>());
+        server server(io_service, tcp::endpoint(tcp::v4(), vm["lp"].as<unsigned short>()), vm["ra"].as<string>(), vm["rp"].as<unsigned short>());
         io_service.run();
     }
     catch (std::exception& e)
